Day9 completed

+++ Day10/Day10Solver.cs
@@ -4,12 +4,12 @@ namespace Day10
 {
     public class Day10Solver : SolverReadAllLines
     {
-        public override int? SolvePart1(string[] input)
+        public override long? SolvePart1(string[] input)
         {
             return base.SolvePart1(input);
         }
 
-        public override int? SolvePart2(string[] input)
+        public override long? SolvePart2(string[] input)
         {
             return base.SolvePart2(input);
         }
